+++ pageSize.h
@@ -1,6 +1,6 @@
 #ifndef PAGESIZE_H
 #define PAGESIZE_H
 
-int init_page_size(void);
+size_t init_page_size(void);
 
 #endif
